+++ src/icemac/addressbook/browser/search/result/handler/base.py
@@ -6,11 +6,10 @@ import zope.interface
 
 @zope.interface.implementer(
     icemac.addressbook.browser.interfaces.IAddressBookBackground)
-class Base(object):
+class Base(icemac.addressbook.browser.base.BaseView):
     """Base for search result handlers."""
 
     @property
     def persons(self):
         addressbook = icemac.addressbook.interfaces.IAddressBook(None)
-        session = icemac.addressbook.browser.base.get_session(self.request)
-        return [addressbook[id] for id in session['person_ids']]
+        return [addressbook[id] for id in self.session['person_ids']]